test(scoring): split scoring-context tests into dedicated module
Addresses CodeRabbit review feedback on ace-step/ace-step-1.5#1097:

1. 200 LOC cap (nitpick).
   Previous commit pushed ``scoring_test.py`` to 252 LOC, above the
   project's 200 LOC hard cap.  Move ``LoadScoringModelContextTests``
   (plus its ``_FakeModel`` and ``_FakeHandler`` helpers) into a
   dedicated module ``lm_score_context_test.py``.  Post-split:
   - ``scoring_test.py``:              147 LOC
   - ``lm_score_context_test.py``:     150 LOC

2. Missing docstrings on helper classes/methods.
   ``_FakeModel``, ``_FakeHandler`` and their methods were
   originally inner classes of the test case and lacked docstrings.
   Pull them out to module level in the new file and document each
   class and method per project style (purpose, params, side-effects,
   return value).

Both test modules run cleanly together (23 tests total, all green).

Refs: #1081
