+++ acestep/core/scoring/scoring_test.py
@@ -11,7 +11,6 @@ from acestep.core.scoring.lm_score import (
     pmi_score,
     pmi_to_normalized_score,
     calculate_reward_score,
-    _load_scoring_model_context,
 )
 
 
@@ -139,113 +138,9 @@ class RewardScoreTests(unittest.TestCase):
         self.assertAlmostEqual(total, 0.75, places=2)
 
 
-class LoadScoringModelContextTests(unittest.TestCase):
-    """Regression tests for the Autoscore lifecycle contract (issue #1081).
-
-    These assert three invariants that protect against the MLX unified-
-    memory leak:
-
-    1. ``_load_scoring_model_context`` performs exactly one load/offload
-       cycle per outermost entry, regardless of how deeply nested the
-       inner re-entries are.
-    2. Nested re-entries from the same thread are no-ops and do not
-       trigger additional CPU↔accelerator migrations.
-    3. On the MLX backend with ``offload_to_cpu=True`` the cached HF
-       scoring model is released on outermost exit so the ~8 GB duplicate
-       PyTorch copy does not remain resident between generations.
-    """
-
-    class _FakeModel:
-        def __init__(self):
-            self.device_calls = []
-
-        def to(self, device):
-            self.device_calls.append(str(device))
-            return self
-
-    class _FakeHandler:
-        def __init__(self, backend, offload=True):
-            self.llm_backend = backend
-            self.llm_initialized = True
-            self.offload_to_cpu = offload
-            self.device = "cuda"  # synthetic; _FakeModel.to is a recorder
-            self._hf_model_for_scoring = LoadScoringModelContextTests._FakeModel()
-            self.get_calls = 0
-
-        def get_hf_model_for_scoring(self):
-            self.get_calls += 1
-            return self._hf_model_for_scoring
-
-    def test_single_load_offload_per_outer_context_mlx(self):
-        """One outer entry should trigger exactly one load and one offload."""
-        handler = self._FakeHandler("mlx")
-        model = handler._hf_model_for_scoring
-        with _load_scoring_model_context(handler):
-            pass
-        # Exactly one load (to accelerator) and one offload (to cpu).
-        self.assertEqual(
-            model.device_calls, ["cuda", "cpu"],
-            "expected exactly one load+offload cycle",
-        )
-
-    def test_nested_entries_are_noops_mlx(self):
-        """Nested re-entries must not move the model again."""
-        handler = self._FakeHandler("mlx")
-        model = handler._hf_model_for_scoring
-        with _load_scoring_model_context(handler):
-            calls_after_outer_load = list(model.device_calls)
-            # Deep nesting (simulating _get_logits_and_target_for_scoring
-            # called many times per Autoscore pass).
-            with _load_scoring_model_context(handler):
-                with _load_scoring_model_context(handler):
-                    # Inner contexts must not migrate the model again.
-                    self.assertEqual(model.device_calls, calls_after_outer_load)
-        # After the outermost exit the handler has dropped the cached model,
-        # so use ``model`` (the captured reference) to verify the offload.
-        self.assertEqual(
-            model.device_calls, ["cuda", "cpu"],
-            "nested entries should not add extra migrations",
-        )
-
-    def test_mlx_outermost_exit_drops_cached_model(self):
-        """On MLX+offload, outer exit must clear ``_hf_model_for_scoring``."""
-        handler = self._FakeHandler("mlx", offload=True)
-        self.assertIsNotNone(handler._hf_model_for_scoring)
-        with _load_scoring_model_context(handler):
-            # Still cached while we're inside the context.
-            self.assertIsNotNone(handler._hf_model_for_scoring)
-        # Released after outermost exit so unified memory is returned to OS.
-        self.assertIsNone(handler._hf_model_for_scoring)
-
-    def test_vllm_outermost_exit_keeps_cached_model(self):
-        """vllm backend must NOT drop the cached HF model (CUDA is fine)."""
-        handler = self._FakeHandler("vllm", offload=True)
-        cached = handler._hf_model_for_scoring
-        with _load_scoring_model_context(handler):
-            pass
-        # vllm keeps the cached HF scoring model between passes; the MLX-
-        # specific release path must not fire here.
-        self.assertIs(handler._hf_model_for_scoring, cached)
-
-    def test_mlx_no_offload_keeps_cached_model(self):
-        """Without offload_to_cpu the MLX release path must not fire."""
-        handler = self._FakeHandler("mlx", offload=False)
-        cached = handler._hf_model_for_scoring
-        with _load_scoring_model_context(handler):
-            pass
-        # Not offloading means no load/offload transitions and no drop.
-        self.assertIs(handler._hf_model_for_scoring, cached)
-        self.assertEqual(cached.device_calls, [])
-
-    def test_get_hf_called_only_by_outermost_entry(self):
-        """Nested entries must not re-query ``get_hf_model_for_scoring``."""
-        handler = self._FakeHandler("mlx")
-        with _load_scoring_model_context(handler):
-            with _load_scoring_model_context(handler):
-                with _load_scoring_model_context(handler):
-                    pass
-        # Outer entry queries once; nested entries are pure no-ops.
-        self.assertEqual(handler.get_calls, 1)
+# Regression tests for ``_load_scoring_model_context`` (issue #1081) live in
+# ``acestep/core/scoring/lm_score_context_test.py`` to keep this file under
+# the project's 200 LOC hard cap.
 
 
 if __name__ == "__main__":
